refactor: extract routerMode into a component-free module
features/data was taking the codebase's first runtime dependency on a
component module; Meta/Redirects now import from router-mode.ts too.
Also document the load-bearing declaration order of the ref-sync effect.
