+++ src/components/Router/router-mode.ts
@@ -1,0 +1,1 @@
+export const routerMode = import.meta.env.VITE_ROUTER === 'browser' ? 'browser' : 'hash';
